+++ StreamChatSwiftUITests/Tests/ChatChannel/ChannelInfo/MemberAddView_Tests.swift
@@ -10,31 +10,31 @@ import StreamSwiftTestHelpers
 import SwiftUI
 import XCTest
 
-@MainActor class AddUsersView_Tests: StreamChatTestCase {
-    func test_addUsersView_snapshot() {
+@MainActor class MemberAddView_Tests: StreamChatTestCase {
+    func test_memberAddView_snapshot() {
         // Given
         let searchController = ChatUserSearchController_Mock.mock(client: chatClient)
         let users = ChannelInfoMockUtils.generateMockUsers(count: 20)
         searchController.users_mock = users
-        let viewModel = AddUsersViewModel(
+        let viewModel = MemberAddViewModel(
             loadedUserIds: [],
             searchController: searchController
         )
 
         // When
-        let view = AddUsersView(viewModel: viewModel, onConfirm: { _ in })
+        let view = MemberAddView(factory: DefaultTestViewFactory.shared, viewModel: viewModel, onConfirm: { _ in })
             .applyDefaultSize()
 
         // Then
-        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+        AssertSnapshot(view)
     }
 
-    func test_addUsersView_selectedUsersSnapshot() {
+    func test_memberAddView_selectedUsersSnapshot() {
         // Given
         let searchController = ChatUserSearchController_Mock.mock(client: chatClient)
         let users = ChannelInfoMockUtils.generateMockUsers(count: 10)
         searchController.users_mock = users
-        let viewModel = AddUsersViewModel(
+        let viewModel = MemberAddViewModel(
             loadedUserIds: [],
             searchController: searchController
         )
@@ -42,29 +42,29 @@ import XCTest
         viewModel.toggleUser(users[3])
 
         // When
-        let view = AddUsersView(viewModel: viewModel, onConfirm: { _ in })
+        let view = MemberAddView(factory: DefaultTestViewFactory.shared, viewModel: viewModel, onConfirm: { _ in })
             .applyDefaultSize()
 
         // Then
-        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+        AssertSnapshot(view)
     }
 
-    func test_addUsersView_alreadyMemberSnapshot() {
+    func test_memberAddView_alreadyMemberSnapshot() {
         // Given
         let searchController = ChatUserSearchController_Mock.mock(client: chatClient)
         let users = ChannelInfoMockUtils.generateMockUsers(count: 10)
         searchController.users_mock = users
         let alreadyMemberIds = Array(users.prefix(3)).map(\.id)
-        let viewModel = AddUsersViewModel(
+        let viewModel = MemberAddViewModel(
             loadedUserIds: alreadyMemberIds,
             searchController: searchController
         )
 
         // When
-        let view = AddUsersView(viewModel: viewModel, onConfirm: { _ in })
+        let view = MemberAddView(factory: DefaultTestViewFactory.shared, viewModel: viewModel, onConfirm: { _ in })
             .applyDefaultSize()
 
         // Then
-        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+        AssertSnapshot(view)
     }
 }
